tests/vhdl/requirements.txt
# VUnit VHDL Test Requirements
# Install with: pip3 install -r requirements.txt
vunit-hdl>=4.7.0
